:hammer: Refactor the project to clarify dao and dto
