+++ composeApp/src/commonMain/kotlin/com/clipevery/dao/SyncDao.kt
@@ -1,7 +1,7 @@
 package com.clipevery.dao
 
 import com.clipevery.Database
-import com.clipevery.model.sync.SyncInfo
+import com.clipevery.dto.model.SyncInfo
 
 interface SyncDao {
